requirements.txt
numpy>=1.26.0
pandas>=2.2.0
torch>=2.4.0
matplotlib>=3.9.0
pyyaml>=6.0.1